Module adview_manager::manager

The AdView Manager

Structs

HistoryEntry

History entry of impressions (won auctions) which the Manager holds.

Manager

The AdView Manager

NextAdUnit

The next AdUnit to be shown

Options

The Ad Manager’s options for showing ads.

Size

AdSlot size width x height in pixels (px)

StickyAdUnit

A sticky AdUnit.

Enums

Error

Statics

DEFAULT_MARKET_URL

Default Market Url that can be used for the Options.

DEFAULT_TOKENS

Default Whitelisted token Addresses that can be used for the Options.